Fixing version check
