+++ cmd/main.go
@@ -1,0 +1,7 @@
+package main
+
+var version string
+
+func main() {
+	Execute()
+}
