--- main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/perriea/tfversion/cmd"
-
-func main() {
-	cmd.Execute()
-}
