--- version/version.go
@@ -1,36 +1,0 @@
-package version
-
-import (
-	"fmt"
-
-	version "github.com/hashicorp/go-version"
-)
-
-// Version number that is being run at the moment.
-const Version = "0.1.6"
-
-// Repository Git
-const Repository = "https://github.com/perriea/tfversion"
-
-// Prerelease marker for the version. If this is "" (empty string)
-// then it means that it is a final release. Otherwise, this is a pre-release
-// such as "dev" (in development), "beta", "rc1", etc.
-var Prerelease = ""
-
-// SemVer is an instance of version.Version. This has the secondary
-// benefit of verifying during tests and init time that our version is a
-// proper semantic version, which should always be the case.
-var SemVer *version.Version
-
-func init() {
-	SemVer = version.Must(version.NewVersion(Version))
-}
-
-// String returns the complete version string, including prerelease
-func String() string {
-	if Prerelease != "" {
-		return fmt.Sprintf("%s-%s", Version, Prerelease)
-	}
-
-	return Version
-}
